backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf
python-multipart
openai>=1.54.0
python-dotenv>=1.0.0
